Add a pre-analyzer which determines the size of locals for local.get/set/tee.
The plan is to use this pre-analyzer to do other tasks too, such as performing
stack analysis and determining branch targets. Admittedly, it's a little
unfortunate that we have to parse all the instructions again - context registers
can't be set after the instructions are parsed - but luckily all the parsing is
in this commit so we should be able to perform the necessary analysis inline.