Simply copy data segments into memory rather than creating blocks.
While it's nice to be able to see the file bytes for each initialized block in
memory, the use of separate memory blocks means that you cannot create data
types that cross block boundaries, making some binaries hard to analyze (e.g.
ones in which runs of zeros - even in the middle of structures - are "optimized
out" by creating multiple data segments). In order to support this copying, also
initialize all of memory to 0.