Dynamically create WasmFunctionPreAnalysis objects on demand.
This allows things to work when the DB is reopened (previously, PCode injection
would simply break - oops). Also, take this opportunity to have the preanalyzer
do the disassembly itself. This has the advantage of disassembling everything,
instead of having the disassembler heuristically decide not to disassemble certain
operations (such as long repeats of identical br_table targets or unreachable
instructions).