Use two mem blocks to cover all imports and functions, instead of one per function.
We can get away with doing this because our pre-analyzer dictates exactly what
to disassemble, and this massively speeds up working with large files with many
functions. One sample, with 20+ MB of code and 70000+ functions, went from over
30 minutes for initial loading + 5 minutes for opening the database (each time!)
to less than 2 minutes for loading + less than 10 seconds for opening the
database.