Speed up getFunctionType by directly getting the type.
We were previously calling getNonImportedFunctions just to get a single type
index; since getNonImportedFunctions is not cached, this would regenerate the
entire list each call. This massively slowed down creation of WasmAnalysis
objects.

For now, getNonImportedFunctions is not used anywhere, so just delete it. If
it proves useful in the future, a cached version should be developed to avoid
creating a new list every time.