Introduce a StructureBuilder class to build datatypes easily.
It turns out that adding fields one-at-a-time to StructureDataType produces
quadratic behaviour: the add function calls repack at the end, which walks over
all of the existing fields. To alleviate this, introduce our own
StructureDataType subclass which skips the repack, and insert all the added
fields into a regular StructureDataType at the end to get proper packing
behaviour. This also simplifies the creation of structures further.