spectest bugfix: initialize tables and element segments with 0x00
Tables and element segments apparently default to holding null refs, and since
we're using 0x00000000 for null (`ref.null`, `ref.is_null`), we need to ensure
that they are initialized appropriately.

In the future, we may want to revisit whether zero is the right value for null -
e.g. can zero be a valid value for some reference type?