Call AutoAnalysisManager.codeDefined after disassembling.
We need to call codeDefined after defining any code blocks in order to inform
the auto-analysis manager of the new code, since we're invoking the disassembler
manually.

This allows the auto-analysis manager to correctly trigger the Basic Constant
Reference Analyzer, among others, which enables correct inference of operand
references in many cases. This in turn drastically improves the usability of the
resulting disassembly and decompilation.

Fixes #26.